revert grapheme-counter.js separated entry
