--- benchmark/grapheme/bundle-entries/unicode-segmenter-counter.js
@@ -1,3 +1,0 @@
-export {
-  countGraphemes,
-} from '../../../src/grapheme-counter.js';
