+++ benchmark/grapheme/bundle-entries/unicode-segmenter-full.js
@@ -1,0 +1,1 @@
+export * from '../../../src/grapheme.js';
