+++ benchmark/grapheme/bundle-entries/unicode-segmenter.js
@@ -1,5 +1,3 @@
 export {
   graphemeSegments,
-  splitGraphemes,
-  countGraphemes,
 } from '../../../src/grapheme.js';
