+++ benchmark/grapheme/bundle-stats-hermes.js
@@ -18,8 +18,7 @@ let config = await Metro.loadConfig(undefined, {
 
 let libs = [
   ['unicode-segmenter/grapheme', 'bundle-entries/unicode-segmenter.js'],
-  ['unicode-segmenter/grapheme-counter', 'bundle-entries/unicode-segmenter-counter.js'],
-  ['unicode-segmenter/grapheme + grapheme-counter', 'bundle-entries/unicode-segmenter-both.js'],
+  ['unicode-segmenter/grapheme (full)', 'bundle-entries/unicode-segmenter-full.js'],
   ['graphemer', 'bundle-entries/graphemer.js'],
   ['grapheme-splitter', 'bundle-entries/grapheme-splitter.js'],
   ['@formatjs/intl-segmenter', 'bundle-entries/formatjs-intl-segmenter.js'],
