+++ benchmark/grapheme/bundle-stats.js
@@ -10,8 +10,7 @@ let baseDir = import.meta.dirname;
 
 let libs = [
   ['unicode-segmenter/grapheme', 'bundle-entries/unicode-segmenter.js'],
-  ['unicode-segmenter/grapheme-counter', 'bundle-entries/unicode-segmenter-counter.js'],
-  ['unicode-segmenter/grapheme + grapheme-counter', 'bundle-entries/unicode-segmenter-both.js'],
+  ['unicode-segmenter/grapheme (full)', 'bundle-entries/unicode-segmenter-full.js'],
   ['graphemer', 'bundle-entries/graphemer.js'],
   ['grapheme-splitter', 'bundle-entries/grapheme-splitter.js'],
   ['@formatjs/intl-segmenter', 'bundle-entries/formatjs-intl-segmenter.js'],
