+++ package.json
@@ -30,7 +30,6 @@
     "./emoji": "./src/emoji.js",
     "./general": "./src/general.js",
     "./grapheme": "./src/grapheme.js",
-    "./grapheme-counter": "./src/grapheme-counter.js",
     "./utils": "./src/utils.js",
     "./intl-adapter": {
       "types": "./src/intl-adapter.js",
@@ -66,10 +65,6 @@
       "./grapheme": {
         "import": "./grapheme.js",
         "require": "./grapheme.cjs"
-      },
-      "./grapheme-counter": {
-        "import": "./grapheme-counter.js",
-        "require": "./grapheme-counter.cjs"
       },
       "./utils": {
         "import": "./utils.js",
