+++ test/corpora.js
@@ -9,9 +9,6 @@ import {
   splitGraphemes,
   countGraphemes,
 } from 'unicode-segmenter/grapheme';
-import {
-  countGraphemes as countGraphemesFast,
-} from 'unicode-segmenter/grapheme-counter';
 
 // Real-world text fixtures generated by `scripts/corpora.js`.
 //
@@ -82,7 +79,6 @@ test('corpora', async t => {
 
       // Bind the other public APIs to the same corpus.
       assert.equal(countGraphemes(text), expected.length);
-      assert.equal(countGraphemesFast(text), expected.length);
 
       let split = [...splitGraphemes(text)];
       assert.equal(split.length, expected.length);
